LayoutTests/fast/xmlhttprequest/xmlhttprequest-sync-disabled-expected.txt

CONSOLE MESSAGE: line 1: Synchronous XMLHttpRequests are disabled for this page.
This tests that synchronous XMLHttpRequests fail when they are disabled for documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhr.open("GET", "http://mydomain/", false) threw exception InvalidAccessError (DOM Exception 15): The object does not support the operation or argument..
PASS successfullyParsed is true

TEST COMPLETE
